Fix creating scaled.scale and scaled.center
R attr()s do not get preserved when assigning data into a sub-range of a larger
frame (even if the range is just "whole data"); for that reason attr(fsom$data,
"scaled:scale") is not created after a recent update of scale handing. This
attempts to reconstruct a reasonable scaled.scale/scaled.center in both new
cases.